add image Api as methdod in commentApi and formationApi

+++ src/main/java/com/cinema/minute/ui/Model/Response/FormationResponse/formationResponse.java
@@ -28,5 +28,5 @@ public class formationResponse {
     private LocalDate FinalDate;
     private String description;
     private List<ChapterResponse> chapter;
-    private double durationPerNow;
+    private long durationPerNow;
 }
